billiard walk for v-polytopes: 1. get the vertices of the facet 2. construct the facet 3. compute the reflection.
